audio_transcriber/requirements.txt
torch
torchaudio
transformers
pyannote.audio
whisperx
faster-whisper
numpy
scipy
pydub
